increase training_iters in demo toy problem
A friend an I found that 20 iterations wasn't enough to converge on a solution. Increasing to 50 iterations successfully executed on my old laptop and resulted in decent prediction accuracy.

I would suggest increasing the iterations in the toy problem because it is very important that this example works out of the box.